+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Added
+- Custom OC rule for external clinical data.
+
 [1.0.3] - 2017-01-03
 ---------------------
 ##### Changed